Refactor: 이벤트 분리
