+++ src/pages/LoginPage.jsx
@@ -1,12 +1,22 @@
 /** @jsx createVNode */
 import { createVNode } from '../lib';
+import { userStorage } from '../storages';
+import { globalStore } from '../stores';
 
 export const LoginPage = () => {
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const $username = document.querySelector('#username');
+    const userData = { username: $username.value, email: '', bio: '' };
+    userStorage.set(userData);
+    globalStore.setState({ currentUser: userData, loggedIn: true });
+  };
+
   return (
     <div className='bg-gray-100 flex items-center justify-center min-h-screen'>
       <div className='bg-white p-8 rounded-lg shadow-md w-full max-w-md'>
         <h1 className='text-2xl font-bold text-center text-blue-600 mb-8'>항해플러스</h1>
-        <form id='login-form'>
+        <form id='login-form' onSubmit={handleSubmit}>
           <input
             type='text'
             id='username'
