+++ src/pages/ProfilePage.jsx
@@ -2,10 +2,21 @@
 import { createVNode } from '../lib';
 import { globalStore } from '../stores';
 import { Header, Navigation, Footer } from '../components';
+import { userStorage } from '../storages';
 
 export const ProfilePage = () => {
   const { loggedIn, currentUser } = globalStore.getState();
   const { username, email, bio } = currentUser ?? {};
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const $username = document.querySelector('#username');
+    const $email = document.querySelector('#email');
+    const $bio = document.querySelector('#bio');
+    const userData = { username: $username.value, email: $email.value, bio: $bio.value };
+    userStorage.set(userData);
+    globalStore.setState({ currentUser: userData });
+  };
 
   return (
     <div className='bg-gray-100 min-h-screen flex justify-center'>
@@ -15,7 +26,7 @@ export const ProfilePage = () => {
         <main className='p-4'>
           <div className='bg-white p-8 rounded-lg shadow-md'>
             <h2 className='text-2xl font-bold text-center text-blue-600 mb-8'>내 프로필</h2>
-            <form id='profile-form'>
+            <form id='profile-form' onSubmit={handleSubmit}>
               <div className='mb-4'>
                 <label for='username' className='block text-gray-700 text-sm font-bold mb-2'>
                   사용자 이름
